android/app/src/main/java/com/vava/mapp/AudioManager.kt
package com.vava.mapp

import android.content.Context
import android.media.AudioManager

class AudioManager(private val context: Context) {

    private val audioManager: AudioManager = context.getSystemService(Context.AUDIO_SERVICE) as AudioManager

    fun setOutputDevice(deviceName: String) {
        // Logique pour définir le périphérique de sortie audio
    }

    fun createVirtualOutput(virtualOutputName: String) {
        // Logique pour créer une sortie virtuelle
    }

    fun getCurrentOutputDevice(): String {
        // Retourne le nom du périphérique de sortie actuel
        return audioManager.getProperty(AudioManager.PROPERTY_OUTPUT_SAMPLE_RATE) ?: "default"
    }

    fun startSystemAudio() {
        // Logique pour démarrer le son système
    }

    fun stopSystemAudio() {
        // Logique pour arrêter le son système
    }
}
